requirements.txt
Flask
gunicorn
sqlalchemy
pymysql
python-decouple